+++ cjs-wrapper.js
@@ -1,1 +1,4 @@
-module.exports = require("./dist/index.js").default
+module.exports = {
+  ...require("./dist/index.js").default,
+  ...require("./dist/index.js"),
+}
