+++ package.json
@@ -42,16 +42,19 @@
     "node-inspect-extracted": "1.1.0",
     "ora": "5.4.1",
     "p-retry": "4.6.1",
+    "svix": "0.54.2",
     "typed-emitter": "2.1.0",
     "yargs": "17.3.1"
   },
   "devDependencies": {
+    "@ava/get-port": "1.0.0",
     "@semantic-release/commit-analyzer": "9.0.2",
     "@semantic-release/git": "10.0.1",
     "@semantic-release/npm": "9.0.0",
     "@semantic-release/release-notes-generator": "10.0.3",
     "@swc/core": "1.2.133",
     "@types/lodash": "^4.14.179",
+    "@types/ms": "0.7.31",
     "@types/node": "17.0.10",
     "@types/uuid": "8.3.4",
     "@types/yargs": "17.0.8",
@@ -63,9 +66,12 @@
     "esbuild-runner": "2.2.1",
     "get-port": "6.1.2",
     "husky": ">=6",
+    "internal-ip": "7.0.0",
     "knex": "1.0.2",
     "lint-staged": "12.3.1",
+    "ms": "2.1.3",
     "nanoid": "3.2.0",
+    "p-event": "4.2.0",
     "pgknexlove": "1.1.21",
     "pkg": "5.5.2",
     "playwright": "1.19.2",
